Separate incremental and non-incremental skip-when-empty steps
We don't need to do cleanup in the non-incremental version.
